Frontend file structure refactor
