+++ src/frontend-pwa/src/components/commonComponents/Footer/Footer.tsx
