+++ src/frontend-pwa/src/components/commonComponents/SearchBar/SearchBar.tsx
