+++ src/frontend-pwa/src/components/commonComponents/SearchBar/searchbar.styles.ts
@@ -4,7 +4,7 @@
  */
 
 import styled from '@emotion/styled';
-import typography from '../../typography';
+import typography from '../../../typography';
 
 type SearchBarProps = {
   borderRadius: boolean;
