+++ src/frontend-pwa/src/components/commonComponents/Slider/Slider.tsx
