+++ src/frontend-pwa/src/components/commonComponents/Spinner/spinner.styles.ts
