+++ src/frontend-pwa/src/components/commonComponents/Toggle/toggle.styles.ts
