+++ src/frontend-pwa/src/components/listComponents/LocationListItem/locationlistitem.styles.ts
@@ -3,7 +3,7 @@
  * @author Dallas Richmond
  */
 import styled from '@emotion/styled';
-import typography from '../../typography';
+import typography from '../../../typography';
 
 export const TableData = styled.td`
  ${typography.toString()}
