+++ src/frontend-pwa/src/views/Location/Location.tsx
@@ -10,12 +10,12 @@
  * @author Dallas Richmond, LocalNewsTV
  */
 import { useState } from 'react';
-import ListItems from '../../components/ListItems/ListItems';
+import ListItems from '../../components/listComponents/ListItems/ListItems';
 import SingleLocation from '../../Type/SingleLocation';
 import useAppService from '../../services/app/useAppService';
-import LocationListItem from '../../components/LocationListItem/LocationListItem';
-import SearchBar from '../../components/SearchBar/SearchBar';
-import Mapping from '../../components/Mapping/Mapping';
+import LocationListItem from '../../components/listComponents/LocationListItem/LocationListItem';
+import SearchBar from '../../components/commonComponents/SearchBar/SearchBar';
+import Mapping from '../../components/utilityComponents/Mapping/Mapping';
 import CalcDistance from '../../utils/CalcDistance';
 import { localStorageKeyExists } from '../../utils/AppLocalStorage';
 import constants from '../../constants/Constants';
